Moved PomCustomizer to a better package
